Move PhraseReplacer into vocabulary.py to avoid relative import
